strapi: add law-soc Strapi single-type
A single-type is a model that has only one instance, which is perfect for what
we want since we want to configure views (vs collections, which imply multiple
copies). We could have a "TWC2" type and a "LawSoc" type, but most likely we'll
break them down by scenes e.g. each page gets its own single type.

The shape of the API response is dependent on what we configure in Strapi, which
means that end users shouldn't be able to modify that, only devs. Otherwise, we
could catastrophically break the frontend.

Also, changes should be versioned, otherwise we won't know if we'll accidentally
end up breaking the frontend. I wonder if there's a way to generate checks like
this automatically. Probably not without GraphQL. Not a big deal though, since
we can easily catch this in code review.

I wonder how we should update the Strapi configuration in prod. It looks like
the model configuration is separate from the actual db, so it might be as simple
as pushing the new config files to the prod service.

Configuration notes:

- Enable unauthenticated users to see all data by default. Nothing is private
  anyway.
- Disable draft + publish: It's a PITA. You need to click "publish", otherwise
  the changes don't become live. And "publish" has a check mark next to it,
  which makes you think the changes are already live. Ugh.